requirements.txt
Flask
requests
google-generativeai
markdown2
